Struct rbml::io::SeekableMemWriterUnstable

pub struct SeekableMemWriter {
    // some fields omitted
}

Writes to an owned, growable byte vector that supports seeking.

Example
extern crate rbml; fn main() { #![allow(unused_must_use)] use rbml::io::SeekableMemWriter; let mut w = SeekableMemWriter::new(); w.write(&[0, 1, 2]); assert_eq!(w.unwrap(), [0, 1, 2]); }

use rbml::io::SeekableMemWriter;

let mut w = SeekableMemWriter::new();
w.write(&[0, 1, 2]);

assert_eq!(w.unwrap(), [0, 1, 2]);

Methods

impl SeekableMemWriter

fn new() -> SeekableMemWriter

Create a new SeekableMemWriter.

fn with_capacity(n: usize) -> SeekableMemWriter

Create a new SeekableMemWriter, allocating at least n bytes for the internal buffer.

fn get_ref<'a>(&'a self) -> &'a [u8]

Acquires an immutable reference to the underlying buffer of this SeekableMemWriter.

No method is exposed for acquiring a mutable reference to the buffer because it could corrupt the state of this MemWriter.

fn unwrap(self) -> Vec<u8>

Unwraps this SeekableMemWriter, returning the underlying buffer